Update on entity types. Added immutable types (City, County etc.). Work in progress: DO NOT DEPLOY.
git-svn-id: https://wush.net/svn/mindalliance@1622 baad322d-9929-0410-88f0-f92e8ff3e1bd

+++ channels/trunk/src/main/java/com/mindalliance/channels/command/DefaultCommander.java
@@ -362,7 +362,7 @@ public class DefaultCommander extends AbstractService implements Commander, Init
             if ( name != null && !name.trim().isEmpty() ) {
                 ModelObject mo = queryService.getDao().find( clazz, name.trim() );
                 if ( mo != null && !mo.isUnknown() && mo.isUndefined() ) {
-                    if ( !queryService.isReferenced( mo ) ) {
+                    if ( !queryService.isReferenced( mo ) && !mo.isImmutable() ) {
                         LOG.info( "Removing unused " + mo.getClass().getSimpleName() + " " + mo );
                         queryService.remove( mo );
                     }
